Cleaned up the code & comments. (Sorry about the last couple commits...apparently my HEAD wasn't on the master.)
I moved a lot of the #defines to their own files since we won't be looking or messing with them. At all.

I also removed some of the pointless comments and redid the header so that it is collapsable in the IDE.

BTW, "jan" stands for "Jon, Alex, and Nic". The order was only because it made a pronouncable word...we can always switch to "naj".